lightsensor.py
import sys
import time
import logging
import threading
import RPi.GPIO as GPIO
import requests
import smtplib
import json
import builtins
from time import localtime, strftime

from ConfigParser import SafeConfigParser

